Add types to respective files
