+++ api/openapi-spec/swagger.json
@@ -1939,7 +1939,7 @@
           "type": "string"
         },
         "pluginName": {
-          "description": "Specifies the name of the plugin to use for this connection. Default plugin:\n\t- for postgres: postgresql-database-plugin\n - for mysql: mysql-database-plugin",
+          "description": "Specifies the name of the plugin to use for this connection. Default plugin:\n\t- for postgres: postgresql-database-plugin",
           "type": "string"
         }
       }
