+++ apis/authorization/v1alpha1/crds.yaml
@@ -359,8 +359,7 @@ spec:
                   type: string
                 pluginName:
                   description: "Specifies the name of the plugin to use for this connection.
-                    Default plugin:\n\t- for postgres: postgresql-database-plugin\n
-                    - for mysql: mysql-database-plugin"
+                    Default plugin:\n\t- for postgres: postgresql-database-plugin"
                   type: string
               required:
               - name
