+++ pkg/controller/mysql_role_binding.go
@@ -34,8 +34,12 @@ func (c *UserManagerController) initMysqlRoleBindingWatcher() {
 	c.myRoleBindingInformer = c.dbInformerFactory.Authorization().V1alpha1().MysqlRoleBindings().Informer()
 	c.myRoleBindingQueue = queue.New(api.ResourceKindMysqlRoleBinding, c.MaxNumRequeues, c.NumThreads, c.runMysqlRoleBindingInjector)
 
-	// TODO: add custom event handler?
-	c.myRoleBindingInformer.AddEventHandler(queue.DefaultEventHandler(c.myRoleBindingQueue.GetQueue()))
+	c.myRoleBindingInformer.AddEventHandler(queue.NewEventHandler(c.myRoleBindingQueue.GetQueue(), func(old interface{}, new interface{}) bool {
+		oldObj := old.(*api.MysqlRoleBinding)
+		newObj := new.(*api.MysqlRoleBinding)
+		return newObj.DeletionTimestamp != nil || !newObj.AlreadyObserved(oldObj)
+	}))
+
 	c.myRoleBindingLister = c.dbInformerFactory.Authorization().V1alpha1().MysqlRoleBindings().Lister()
 }
 
@@ -50,7 +54,7 @@ func (c *UserManagerController) runMysqlRoleBindingInjector(key string) error {
 		glog.Warningf("MysqlRoleBinding %s does not exist anymore\n", key)
 
 	} else {
-		mRoleBinding := obj.(*api.MysqlRoleBinding)
+		mRoleBinding := obj.(*api.MysqlRoleBinding).DeepCopy()
 
 		glog.Infof("Sync/Add/Update for MysqlRoleBinding %s/%s\n", mRoleBinding.Namespace, mRoleBinding.Name)
 
@@ -59,17 +63,19 @@ func (c *UserManagerController) runMysqlRoleBindingInjector(key string) error {
 				go c.runMysqlRoleBindingFinalizer(mRoleBinding, 1*time.Minute, 10*time.Second)
 			}
 
-		} else if !kutilcorev1.HasFinalizer(mRoleBinding.ObjectMeta, MysqlRoleBindingFinalizer) {
-			// Add finalizer
-			_, _, err = patchutil.PatchMysqlRoleBinding(c.dbClient.AuthorizationV1alpha1(), mRoleBinding, func(binding *api.MysqlRoleBinding) *api.MysqlRoleBinding {
-				binding.ObjectMeta = kutilcorev1.AddFinalizer(binding.ObjectMeta, MysqlRoleBindingFinalizer)
-				return binding
-			})
-			if err != nil {
-				return errors.Wrapf(err, "failed to set MysqlRoleBinding finalizer for (%s/%s)", mRoleBinding.Namespace, mRoleBinding.Name)
+		} else {
+			if !kutilcorev1.HasFinalizer(mRoleBinding.ObjectMeta, MysqlRoleBindingFinalizer) {
+				// Add finalizer
+				_, _, err = patchutil.PatchMysqlRoleBinding(c.dbClient.AuthorizationV1alpha1(), mRoleBinding, func(binding *api.MysqlRoleBinding) *api.MysqlRoleBinding {
+					binding.ObjectMeta = kutilcorev1.AddFinalizer(binding.ObjectMeta, MysqlRoleBindingFinalizer)
+					return binding
+				})
+				if err != nil {
+					return errors.Wrapf(err, "failed to set MysqlRoleBinding finalizer for (%s/%s)", mRoleBinding.Namespace, mRoleBinding.Name)
+				}
+
 			}
 
-		} else {
 			dbRBClient, err := database.NewDatabaseRoleBindingForMysql(c.kubeClient, c.dbClient, mRoleBinding)
 			if err != nil {
 				return err
