+++ pkg/controller/postgres_role.go
@@ -26,8 +26,12 @@ func (c *UserManagerController) initPostgresRoleWatcher() {
 	c.pgRoleInformer = c.dbInformerFactory.Authorization().V1alpha1().PostgresRoles().Informer()
 	c.pgRoleQueue = queue.New(api.ResourceKindPostgresRole, c.MaxNumRequeues, c.NumThreads, c.runPostgresRoleInjector)
 
-	// TODO: add custom event handler?
-	c.pgRoleInformer.AddEventHandler(queue.DefaultEventHandler(c.pgRoleQueue.GetQueue()))
+	c.pgRoleInformer.AddEventHandler(queue.NewEventHandler(c.pgRoleQueue.GetQueue(), func(old interface{}, new interface{}) bool {
+		oldObj := old.(*api.PostgresRole)
+		newObj := new.(*api.PostgresRole)
+		return newObj.DeletionTimestamp != nil || !newObj.AlreadyObserved(oldObj)
+	}))
+
 	c.pgRoleLister = c.dbInformerFactory.Authorization().V1alpha1().PostgresRoles().Lister()
 }
 
@@ -42,7 +46,7 @@ func (c *UserManagerController) runPostgresRoleInjector(key string) error {
 		glog.Warningf("PostgresRole %s does not exist anymore\n", key)
 
 	} else {
-		pgRole := obj.(*api.PostgresRole)
+		pgRole := obj.(*api.PostgresRole).DeepCopy()
 
 		glog.Infof("Sync/Add/Update for PostgresRole %s/%s\n", pgRole.Namespace, pgRole.Name)
 
@@ -51,16 +55,18 @@ func (c *UserManagerController) runPostgresRoleInjector(key string) error {
 				go c.runPostgresRoleFinalizer(pgRole, 1*time.Minute, 10*time.Second)
 			}
 
-		} else if !kutilcorev1.HasFinalizer(pgRole.ObjectMeta, PostgresRoleFinalizer) {
-			// Add finalizer
-			_, _, err := patchutil.PatchPostgresRole(c.dbClient.AuthorizationV1alpha1(), pgRole, func(role *api.PostgresRole) *api.PostgresRole {
-				role.ObjectMeta = kutilcorev1.AddFinalizer(role.ObjectMeta, PostgresRoleFinalizer)
-				return role
-			})
-			if err != nil {
-				return errors.Wrapf(err, "failed to set postgresRole finalizer for (%s/%s)", pgRole.Namespace, pgRole.Name)
-			}
 		} else {
+			if !kutilcorev1.HasFinalizer(pgRole.ObjectMeta, PostgresRoleFinalizer) {
+				// Add finalizer
+				_, _, err := patchutil.PatchPostgresRole(c.dbClient.AuthorizationV1alpha1(), pgRole, func(role *api.PostgresRole) *api.PostgresRole {
+					role.ObjectMeta = kutilcorev1.AddFinalizer(role.ObjectMeta, PostgresRoleFinalizer)
+					return role
+				})
+				if err != nil {
+					return errors.Wrapf(err, "failed to set postgresRole finalizer for (%s/%s)", pgRole.Namespace, pgRole.Name)
+				}
+			}
+
 			dbRClient, err := database.NewDatabaseRoleForPostgres(c.kubeClient, pgRole)
 			if err != nil {
 				return err
@@ -82,7 +88,7 @@ func (c *UserManagerController) runPostgresRoleInjector(key string) error {
 // 	  - configure a role that maps a name in Vault to an SQL statement to execute to create the database credential.
 //    - sync role
 func (c *UserManagerController) reconcilePostgresRole(dbRClient database.DatabaseRoleInterface, pgRole *api.PostgresRole) error {
-	if pgRole.Status.ObservedGeneration == 0 { // initial stage
+	if pgRole.Status.Phase == "" { // initial stage
 		name := pgRole.Name
 		namespace := pgRole.Namespace
 		status := pgRole.Status
